# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMC Documentation",
        "description": "Api para criaçao e gerenciamento de prontuarios, juntamente com analise de batimentos cardiacos",
        "contact": {
            "url": "https://github.com/gmatheus66"
        },
        "license": {
            "name": "License: MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://apisimc.herokuapp.com/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/doctor/register": {
            "post": {
                "tags": [
                    "Doctor"
                ],
                "description": "Doctor Register",
                "operationId": "App\\Http\\Controllers\\DoctorController::register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true
                    },
                    {
                        "name": "crm",
                        "in": "query",
                        "description": "Crm",
                        "required": true
                    },
                    {
                        "name": "specialization",
                        "in": "query",
                        "description": "Specialization",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/patient/gethandbook": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "description": "Returns patient records",
                "operationId": "App\\Http\\Controllers\\HandbookController::get_handbook_patient",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/doctor/gethandbook": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "description": "Returns doctor records",
                "operationId": "App\\Http\\Controllers\\HandbookController::get_handbook_doctor",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/patient/gethandbook/{id}": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "description": "Returns patient records detail",
                "operationId": "App\\Http\\Controllers\\HandbookController::get_detail_handbook_patient",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/doctor/handbook/register": {
            "post": {
                "tags": [
                    "Doctor"
                ],
                "description": "Handbook Register",
                "operationId": "App\\Http\\Controllers\\HandbookController::register",
                "parameters": [
                    {
                        "name": "name_handbook",
                        "in": "query",
                        "description": "Name Handbook",
                        "required": true
                    },
                    {
                        "name": "limitation",
                        "in": "query",
                        "description": "Limitation in:Cognitive,Locomotion,Vision,Hearing",
                        "required": true
                    },
                    {
                        "name": "body_mass",
                        "in": "query",
                        "description": "Body Mass",
                        "required": true
                    },
                    {
                        "name": "weight",
                        "in": "query",
                        "description": "Weight",
                        "required": true
                    },
                    {
                        "name": "service_date",
                        "in": "query",
                        "description": "Service Date",
                        "required": true
                    },
                    {
                        "name": "complaints",
                        "in": "query",
                        "description": "Complaints",
                        "required": true
                    },
                    {
                        "name": "symptoms",
                        "in": "query",
                        "description": "Symptoms",
                        "required": true
                    },
                    {
                        "name": "vital_signs",
                        "in": "query",
                        "description": "Vital Signs",
                        "required": true
                    },
                    {
                        "name": "blood_type",
                        "in": "query",
                        "description": "Blood Type in:A+,A-,B+,B-,AB+,AB-,O+,O-",
                        "required": true
                    },
                    {
                        "name": "blood_pressure",
                        "in": "query",
                        "description": "Blood Pressure",
                        "required": true
                    },
                    {
                        "name": "hgt",
                        "in": "query",
                        "description": "HGT",
                        "required": true
                    },
                    {
                        "name": "temperature",
                        "in": "query",
                        "description": "Temperature",
                        "required": true
                    },
                    {
                        "name": "relative_id",
                        "in": "query",
                        "description": "Relative ID",
                        "required": true
                    },
                    {
                        "name": "patient_id",
                        "in": "query",
                        "description": "Patient ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}